+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+- Introduced the `Ewald.substrate_n_fold` property to generate rotationally equivalent substrate domains in the Ewald construction.
+
 <a name="2.3.0"></a>
 ## [2.3.0] – 2026-07-29